requirements.txt
dj-database-url==0.4.2
Django==1.11
gunicorn==19.7.1
whitenoise==3.3.0
appdirs==1.4.0
packaging==16.8
pyparsing==2.1.10
six==1.10.0
django-paypal==0.4.1
psycopg2==2.6.2
twilio==6.5.1
